+++ app/Source/Modules/Auth/Adapter/Controller/AuthController.php
@@ -1,0 +1,98 @@
+<?php
+
+namespace App\Source\Modules\Auth\Adapter\Controller;
+
+use App\Source\Modules\Auth\Adapter\Repository\Eloquent\UserModelEloquent;
+use App\Source\Shared\Adapter\Controller\Controller;
+use App\Source\Shared\util\Res;
+use Tymon\JWTAuth\Exceptions\JWTException;
+
+class AuthController extends Controller
+{
+    /**
+     * Create a new AuthController instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth:api', ['except' => ['login']]);
+    }
+
+    /**
+     * Get a JWT via given credentials.
+     *
+     * @return \Illuminate\Http\JsonResponse
+     */
+    public function login()
+    {
+        $credentials = request(['email', 'password']);
+        $token = auth()->attempt($credentials);
+        throw_if(!$token, new JWTException(trans('auth_lang.not_authorized')));
+
+        return $this->respondWithToken($token);
+    }
+
+    /**
+     * Log the user out (Invalidate the token).
+     *
+     * @return \Illuminate\Http\JsonResponse
+     */
+    public function logout()
+    {
+        auth()->logout();
+        return response()->json(['message' => trans('auth_lang.logout')]);
+    }
+
+    /**
+     * Refresh a token.
+     *
+     * @return \Illuminate\Http\JsonResponse
+     */
+    public function refresh()
+    {
+        return $this->respondWithToken(auth()->refresh());
+    }
+
+    /**
+     * Get the authenticated User.
+     *
+     * @return \Illuminate\Http\JsonResponse
+     */
+    public function me()
+    {
+        return Res::success($this->getUserWithPermission());
+    }
+
+    /**
+     * Get the token array structure.
+     *
+     * @param  string $token
+     *
+     * @return \Illuminate\Http\JsonResponse
+     */
+    protected function respondWithToken($token)
+    {
+        $user = $this->getUserWithPermission();
+        // Cache::put($token, $user, (env('JWT_TTL', 240) * 60));
+        return Res::success([
+            'access_token' => $token,
+            'token_type' => 'bearer',
+            'expires_in' => auth('api')->factory()->getTTL() * 60,
+            'user' => $user,
+        ]);
+    }
+
+    /**
+     * Obter dados do usuario logado e retornar permissões de acesso
+     *
+     * @return array
+     */
+    protected function getUserWithPermission(): array
+    {
+        return UserModelEloquent::whereId(auth()->user()->id)
+            // ->with('role.rolePermission')
+            ->first()
+            ->toArray();
+    }
+}
