+++ app/Source/Modules/Stock/Product/Domain/UseCase/ProductStoreUseCase.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace App\Source\Modules\Stock\Product\Domain\UseCase;
+
+use App\Source\Modules\Stock\Product\Adapter\Dto\ProductDto;
+use App\Source\Modules\Stock\Product\Adapter\Mapper\ProductMapper;
+use App\Source\Modules\Stock\Product\Port\Repository\ProductRepositoryInterface;
+use App\Source\Shared\Domain\UseCase\UseCaseBase;
+
+final class ProductStoreUseCase extends UseCaseBase
+{
+    private ProductMapper $mapper;
+    private function __construct(private ProductRepositoryInterface $repository){
+        $this->mapper = ProductMapper::make();
+    }
+
+    public static function make(ProductRepositoryInterface $repository): self {
+        return new self($repository);
+    }
+
+    public function execute(ProductDto $dto): ProductDto {
+        $entityMapped = $this->mapper->dtoToEntity($dto);
+        $entityStored = $this->repository->store($entityMapped);
+        $dtoMapped = $this->mapper->entityToDto($entityStored);
+
+        return $dtoMapped;
+    }
+}
