+++ app/Source/Modules/Stock/Product/Domain/UseCase/ProductUpdateUseCase.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace App\Source\Modules\Stock\Product\Domain\UseCase;
+
+use App\Source\Modules\Stock\Product\Adapter\Dto\ProductDto;
+use App\Source\Modules\Stock\Product\Adapter\Mapper\ProductMapper;
+use App\Source\Modules\Stock\Product\Port\Repository\ProductRepositoryInterface;
+use App\Source\Shared\Domain\UseCase\UseCaseBase;
+
+final class ProductUpdateUseCase extends UseCaseBase
+{
+    private ProductMapper $mapper;
+    private function __construct(private ProductRepositoryInterface $repository){
+        $this->mapper = ProductMapper::make();
+    }
+
+    public static function make(ProductRepositoryInterface $repository): self {
+        return new self($repository);
+    }
+
+    public function execute(ProductDto $dto, int $id): ProductDto {
+        $entityMapped = $this->mapper->dtoToEntity($dto);        
+        $entityUpdated = $this->repository->update($entityMapped, $id);
+        $dtoMapped = $this->mapper->entityToDto($entityUpdated);
+
+        return $dtoMapped;
+    }
+}
