+++ app/Source/Shared/Adapter/Controller/Controller.php
@@ -17,61 +17,5 @@ use Illuminate\Routing\Controller as BaseController;
 
 class Controller extends BaseController
 {
-    use AuthorizesRequests, DispatchesJobs, ValidatesRequests;
-
-    function baseResponse(): array
-    {
-        return [
-            'code' => '',
-            'error' => false,
-            'message' => '',
-            'data' => [],
-        ];
-    }
-
-    public function responseSuccess(mixed $data = [], int $code = Response::HTTP_OK, string $msg = ''): JsonResponse
-    {
-        // Quando nenhuma mensagem informado, seta um default
-        if (!$msg){
-            $msg = match ($code) {
-                Response::HTTP_OK => trans('message_lang.http_ok'),
-                Response::HTTP_CREATED => trans('message_lang.http_created'),
-                Response::HTTP_BAD_REQUEST => trans('message_lang.http_bad_request'),
-                Response::HTTP_NOT_FOUND => trans('message_lang.http_not_found'),
-                default => '',
-            };
-        }
-
-        // Configurar Resposta
-        $baseResponse = $this->baseResponse();
-        $baseResponse['code'] = $code;
-        $baseResponse['error'] = false;
-        $baseResponse['message'] = $msg;
-        $baseResponse['data'] = $data;
-
-        // Retornar Resposta
-        return response()->json($baseResponse, $code);        
-    }
-
-    public function responseError(mixed $data = [], int $code = Response::HTTP_BAD_REQUEST, string $msg = ''): JsonResponse
-    {
-        // Quando nenhuma mensagem informado, seta um default
-        if (!$msg) {
-            $msg = match ($code) {
-                Response::HTTP_BAD_REQUEST => trans('message_lang.http_bad_request'),
-                Response::HTTP_NOT_FOUND => trans('message_lang.http_not_found'),
-                default => '',
-            };
-        }
-
-        // Configurar resposta
-        $baseResponse = $this->baseResponse();
-        $baseResponse['code'] = $code;
-        $baseResponse['error'] = true;
-        $baseResponse['message'] = $msg;
-        $baseResponse['data'] = $data;
-
-        // Retornar Resposta
-        return response()->json($baseResponse, $code);
-    }
+    use AuthorizesRequests, DispatchesJobs, ValidatesRequests;    
 }
