+++ config/auth.php
@@ -66,8 +66,7 @@ return [
     'providers' => [
         'users' => [
             'driver' => 'eloquent',
-            'model' => App\Models\User::class,
-            // 'model' => App\Source\Modules\User\Adapter\Repository\Eloquent\UserModelEloquent::class,
+            'model' => App\Source\Modules\Auth\Adapter\Repository\Eloquent\UserModelEloquent::class,
         ],
 
         // 'users' => [
