+++ routes/api/api.php
@@ -1,31 +1,15 @@
 <?php
 
-use App\Http\Controllers\AuthController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Artisan;
 use Illuminate\Support\Facades\Cache;
 use Illuminate\Support\Facades\Route;
 
-/*
-|--------------------------------------------------------------------------
-| API Routes
-|--------------------------------------------------------------------------
-|
-| Here is where you can register API routes for your application. These
-| routes are loaded by the RouteServiceProvider within a group which
-| is assigned the "api" middleware group. Enjoy building your API!
-|
-*/
-
-Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
-    return $request->user();
-});
-
 /**
- * Teste de conexão com a Central
+ * Teste de conexão
  */
 Route::get('/ping', function () {
-    return 'pong [CENTRAL]';
+    return 'pong';
 });
 
 // Limpar cache
@@ -37,8 +21,3 @@ Route::get('/clear-cache', function () {
     Cache::flush();
     return "Cache is cleared";
 });
-
-Route::post('auth/login', [AuthController::class, 'login']);
-Route::post('logout', 'AuthController@logout');
-Route::post('refresh', 'AuthController@refresh');
-Route::post('me', 'AuthController@me');
